requirements.txt
aiosmtpd==1.4.6
PyYAML==6.0.2
watchdog==6.0.0
